+++ packages/cubejs-query-orchestrator/test/unit/QueryOrchestrator.test.js
@@ -225,7 +225,7 @@ describe('QueryOrchestrator', () => {
   });
 
   test('index is part of query key', async () => {
-    const firstPromise = queryOrchestrator.fetchQuery({
+    queryOrchestrator.fetchQuery({
       query: 'SELECT "orders__created_at_week" "orders__created_at_week", sum("orders__count") "orders__count" FROM (SELECT * FROM stb_pre_aggregations.orders_number_and_count20191102) as partition_union  WHERE ("orders__created_at_week" >= ($1::timestamptz::timestamptz AT TIME ZONE \'UTC\') AND "orders__created_at_week" <= ($2::timestamptz::timestamptz AT TIME ZONE \'UTC\')) GROUP BY 1 ORDER BY 1 ASC LIMIT 10000',
       values: ['2019-11-01T00:00:00Z', '2019-11-30T23:59:59Z'],
       cacheKeyQueries: {